testPredictFStat.sh: control 'cleanup-step' via setting 'NOCLEANUP' environment variable Original: 48f7d3ceefa7f14f89d4cc5196cb6a6c66766f77

+++ lalapps/src/pulsar/FDS_isolated/testPredictFStat.sh
@@ -131,7 +131,9 @@ resPFS2=`echo $tmp | awk '{printf "%g", $1}'`
 
 
 ## clean up files
-rm -rf $SFTdir $outfile_pfs $outfile_pfs2
+if [ -z "$NOCLEANUP" ]; then
+    rm -rf $SFTdir $outfile_pfs $outfile_pfs2;
+fi
 
 echo
 echo "----------------------------------------"
